requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
mfrc522==0.0.7
pydantic==2.11.4
pydantic_core==2.33.2
RPi.GPIO==0.7.1
sniffio==1.3.1
spidev==3.7
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
